# review-b/api/business/v1/business.pb.go
//`protoc --proto_path=./api --proto_path=./third_party --go_out=paths=source_relative:./api --go-http_out=paths=source_relative:./api --go-grpc_out=paths=source_relative:./api --openapi_out=fq_schema_naming=true,default_response=false:. api/business/v1/business.proto`
//protoc --proto_path=. --proto_path=./third_party --go_out=paths=source_relative:. --validate_out=paths=source_relative,lang=go:. ./api/business/v1/business.proto

// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.9
// 	protoc        v6.32.0
// source: api/business/v1/business.proto

package v1

import (
	_ "github.com/envoyproxy/protoc-gen-validate/validate"
	_ "google.golang.org/genproto/googleapis/api/annotations"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type ReplyReviewRequest struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	ReviewID      int64                  `protobuf:"varint,1,opt,name=reviewID,proto3" json:"reviewID,omitempty"`
	StoreID       int64                  `protobuf:"varint,2,opt,name=storeID,proto3" json:"storeID,omitempty"`
	Content       string                 `protobuf:"bytes,3,opt,name=content,proto3" json:"content,omitempty"`
	PicInfo       string                 `protobuf:"bytes,4,opt,name=picInfo,proto3" json:"picInfo,omitempty"`
	VideoInfo     string                 `protobuf:"bytes,5,opt,name=videoInfo,proto3" json:"videoInfo,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *ReplyReviewRequest) Reset() {
	*x = ReplyReviewRequest{}
	mi := &file_api_business_v1_business_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *ReplyReviewRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*ReplyReviewRequest) ProtoMessage() {}

func (x *ReplyReviewRequest) ProtoReflect() protoreflect.Message {
	mi := &file_api_business_v1_business_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use ReplyReviewRequest.ProtoReflect.Descriptor instead.
func (*ReplyReviewRequest) Descriptor() ([]byte, []int) {
	return file_api_business_v1_business_proto_rawDescGZIP(), []int{0}
}

func (x *ReplyReviewRequest) GetReviewID() int64 {
	if x != nil {
		return x.ReviewID
	}
	return 0
}

func (x *ReplyReviewRequest) GetStoreID() int64 {
	if x != nil {
		return x.StoreID
	}
	return 0
}

func (x *ReplyReviewRequest) GetContent() string {
	if x != nil {
		return x.Content
	}
	return ""
}

func (x *ReplyReviewRequest) GetPicInfo() string {
	if x != nil {
		return x.PicInfo
	}
	return ""
}

func (x *ReplyReviewRequest) GetVideoInfo() string {
	if x != nil {
		return x.VideoInfo
	}
	return ""
}

// 回复评价的返回值
type ReplyReviewReply struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	ReplyID       int64                  `protobuf:"varint,1,opt,name=replyID,proto3" json:"replyID,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *ReplyReviewReply) Reset() {
	*x = ReplyReviewReply{}
	mi := &file_api_business_v1_business_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *ReplyReviewReply) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*ReplyReviewReply) ProtoMessage() {}

func (x *ReplyReviewReply) ProtoReflect() protoreflect.Message {
	mi := &file_api_business_v1_business_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use ReplyReviewReply.ProtoReflect.Descriptor instead.
func (*ReplyReviewReply) Descriptor() ([]byte, []int) {
	return file_api_business_v1_business_proto_rawDescGZIP(), []int{1}
}

func (x *ReplyReviewReply) GetReplyID() int64 {
	if x != nil {
		return x.ReplyID
	}
	return 0
}

var File_api_business_v1_business_proto protoreflect.FileDescriptor

const file_api_business_v1_business_proto_rawDesc = "" +
	"\n" +
	"\x1eapi/business/v1/business.proto\x12\x0fapi.business.v1\x1a\x1cgoogle/api/annotations.proto\x1a\x17validate/validate.proto\"\xba\x01\n" +
	"\x12ReplyReviewRequest\x12#\n" +
	"\breviewID\x18\x01 \x01(\x03B\a\xfaB\x04\"\x02 \x00R\breviewID\x12!\n" +
	"\astoreID\x18\x02 \x01(\x03B\a\xfaB\x04\"\x02 \x00R\astoreID\x12$\n" +
	"\acontent\x18\x03 \x01(\tB\n" +
	"\xfaB\ar\x05\x10\x02\x18\xc8\x01R\acontent\x12\x18\n" +
	"\apicInfo\x18\x04 \x01(\tR\apicInfo\x12\x1c\n" +
	"\tvideoInfo\x18\x05 \x01(\tR\tvideoInfo\",\n" +
	"\x10ReplyReviewReply\x12\x18\n" +
	"\areplyID\x18\x01 \x01(\x03R\areplyID2\x86\x01\n" +
	"\bBusiness\x12z\n" +
	"\vReplyReview\x12#.api.business.v1.ReplyReviewRequest\x1a!.api.business.v1.ReplyReviewReply\"#\x82\xd3\xe4\x93\x02\x1d:\x01*\"\x18business/v1/review/replyB0\n" +
	"\x0fapi.business.v1P\x01Z\x1breview-b/api/business/v1;v1b\x06proto3"

var (
	file_api_business_v1_business_proto_rawDescOnce sync.Once
	file_api_business_v1_business_proto_rawDescData []byte
)

func file_api_business_v1_business_proto_rawDescGZIP() []byte {
	file_api_business_v1_business_proto_rawDescOnce.Do(func() {
		file_api_business_v1_business_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_api_business_v1_business_proto_rawDesc), len(file_api_business_v1_business_proto_rawDesc)))
	})
	return file_api_business_v1_business_proto_rawDescData
}

var file_api_business_v1_business_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_api_business_v1_business_proto_goTypes = []any{
	(*ReplyReviewRequest)(nil), // 0: api.business.v1.ReplyReviewRequest
	(*ReplyReviewReply)(nil),   // 1: api.business.v1.ReplyReviewReply
}
var file_api_business_v1_business_proto_depIdxs = []int32{
	0, // 0: api.business.v1.Business.ReplyReview:input_type -> api.business.v1.ReplyReviewRequest
	1, // 1: api.business.v1.Business.ReplyReview:output_type -> api.business.v1.ReplyReviewReply
	1, // [1:2] is the sub-list for method output_type
	0, // [0:1] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_api_business_v1_business_proto_init() }
func file_api_business_v1_business_proto_init() {
	if File_api_business_v1_business_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_api_business_v1_business_proto_rawDesc), len(file_api_business_v1_business_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_api_business_v1_business_proto_goTypes,
		DependencyIndexes: file_api_business_v1_business_proto_depIdxs,
		MessageInfos:      file_api_business_v1_business_proto_msgTypes,
	}.Build()
	File_api_business_v1_business_proto = out.File
	file_api_business_v1_business_proto_goTypes = nil
	file_api_business_v1_business_proto_depIdxs = nil
}